fix(server): use project data for input slots
Resolve input slots from project AST and type information, using the
configured importer and registered classfile packages for candidates.
Limit spx resource analysis to documents registered with the spx framework.

Move general input slots tests and large-list benchmarks to plain XGo and
the isolated classfile framework. Keep sprite, resource, and color
regressions together in dedicated spx tests, and remove the remaining
spx-specific unit test server helper.

Respect declaration visibility and shadowing when collecting and caching
candidates. Cover initializers, nested scopes, cross-file calls, source
kinds, document updates, command dispatch, and incomplete code without
loading embedded package data.
