+++ internal/server/command.go
@@ -7,16 +7,9 @@ import (
 	gotypes "go/types"
 	"iter"
 	"slices"
-	"strconv"
-	"strings"
 	"unicode"
-	"unicode/utf16"
-	"unicode/utf8"
 
-	"github.com/goplus/xgo/ast"
-	"github.com/goplus/xgo/token"
 	"github.com/goplus/xgolsw/pkgdoc"
-	"github.com/goplus/xgolsw/xgo/types"
 	"github.com/goplus/xgolsw/xgo/xgoutil"
 )
 
@@ -56,7 +49,7 @@ func (s *Server) workspaceExecuteCommand(params *ExecuteCommandParams) (any, err
 			}
 			cmdParams = append(cmdParams, cmdParam)
 		}
-		return s.spxGetInputSlots(cmdParams)
+		return s.xgoGetInputSlots(cmdParams)
 	case CommandXGoGetProperties:
 		var cmdParams XGoGetPropertiesParams
 		if len(params.Arguments) != 1 {
@@ -125,27 +118,6 @@ func (s *Server) spxRenameResourcesWithCompileResult(result *compileResult, para
 		}
 	}
 	return &workspaceEdit, nil
-}
-
-// spxGetInputSlots gets input slots in source order. Returned slots have
-// pairwise non-overlapping ranges.
-func (s *Server) spxGetInputSlots(params []XGoGetInputSlotsParams) ([]XGoInputSlot, error) {
-	if l := len(params); l == 0 {
-		return nil, nil
-	} else if l > 1 {
-		return nil, fmt.Errorf("%s only supports one document at a time", CommandXGoGetInputSlots)
-	}
-	param := params[0]
-
-	result, _, astFile, err := s.compileAndGetASTFileForDocumentURI(param.TextDocument.URI)
-	if err != nil {
-		return nil, err
-	}
-	if astFile == nil {
-		return nil, nil
-	}
-
-	return findInputSlots(result, astFile), nil
 }
 
 // xgoGetProperties gets properties for a specific target (e.g., "Game" or a sprite name).
@@ -550,992 +522,4 @@ func findEnclosingType(obj gotypes.Object) *gotypes.Named {
 	}
 
 	return nil
-}
-
-// inputSlotContext caches file-level data shared by all input slots in one
-// request.
-type inputSlotContext struct {
-	result                   *compileResult
-	astFile                  *ast.File
-	astPkg                   *ast.Package
-	typeInfo                 *types.Info
-	parents                  map[ast.Node]ast.Node
-	scopeObjects             map[*gotypes.Scope][]gotypes.Object
-	scopeObjectPositions     map[*gotypes.Scope][]token.Pos
-	predefinedNames          map[predefinedNamesCacheKey][]string
-	utf16ColumnsByByteOffset []uint32
-}
-
-// predefinedNamesCacheKey identifies expressions with the same visible names.
-type predefinedNamesCacheKey struct {
-	scope              *gotypes.Scope
-	declaredType       gotypes.Type
-	visibleObjectCount int
-}
-
-// newInputSlotContext creates a context for finding input slots in astFile.
-func newInputSlotContext(result *compileResult, astFile *ast.File) *inputSlotContext {
-	typeInfo, _ := result.proj.TypeInfo()
-	astPkg, _ := result.proj.ASTPackage()
-	return &inputSlotContext{
-		result:                   result,
-		astFile:                  astFile,
-		astPkg:                   astPkg,
-		typeInfo:                 typeInfo,
-		parents:                  nodeParents(astFile),
-		scopeObjects:             make(map[*gotypes.Scope][]gotypes.Object),
-		scopeObjectPositions:     make(map[*gotypes.Scope][]token.Pos),
-		predefinedNames:          make(map[predefinedNamesCacheKey][]string),
-		utf16ColumnsByByteOffset: buildUTF16ColumnIndex(astFile.Code),
-	}
-}
-
-// nodeParents returns the parent of each descendant of root.
-func nodeParents(root ast.Node) map[ast.Node]ast.Node {
-	parents := make(map[ast.Node]ast.Node)
-	var stack []ast.Node
-	ast.Inspect(root, func(node ast.Node) bool {
-		if node == nil {
-			stack = stack[:len(stack)-1]
-			return true
-		}
-		if len(stack) > 0 {
-			parents[node] = stack[len(stack)-1]
-		}
-		stack = append(stack, node)
-		return true
-	})
-	return parents
-}
-
-// buildUTF16ColumnIndex returns the UTF-16 column at every byte offset in code.
-func buildUTF16ColumnIndex(code []byte) []uint32 {
-	columns := make([]uint32, len(code)+1)
-	var column uint32
-	for offset := 0; offset < len(code); {
-		columns[offset] = column
-		switch code[offset] {
-		case '\n':
-			offset++
-			column = 0
-			columns[offset] = column
-			continue
-		case '\r':
-			if offset+1 < len(code) && code[offset+1] == '\n' {
-				offset++
-				columns[offset] = column
-				continue
-			}
-		}
-
-		r, size := utf8.DecodeRune(code[offset:])
-		for i := 1; i < size; i++ {
-			columns[offset+i] = column
-		}
-		column += uint32(utf16.RuneLen(r))
-		offset += size
-		columns[offset] = column
-	}
-	return columns
-}
-
-// position converts pos to an LSP position using the precomputed UTF-16
-// column index.
-func (c *inputSlotContext) position(pos token.Pos) Position {
-	filePosition := c.result.proj.Fset.Position(pos)
-	offset := min(max(filePosition.Offset, 0), len(c.astFile.Code))
-	return Position{
-		Line:      uint32(max(filePosition.Line, 1) - 1),
-		Character: c.utf16ColumnsByByteOffset[offset],
-	}
-}
-
-// rangeForNode returns the LSP range for node.
-func (c *inputSlotContext) rangeForNode(node ast.Node) Range {
-	return Range{Start: c.position(node.Pos()), End: c.position(node.End())}
-}
-
-// rangeForPosEnd returns the LSP range for pos and end.
-func (c *inputSlotContext) rangeForPosEnd(pos, end token.Pos) Range {
-	return Range{Start: c.position(pos), End: c.position(end)}
-}
-
-// innermostScope returns the innermost type-checker scope containing node.
-func (c *inputSlotContext) innermostScope(node ast.Node) *gotypes.Scope {
-	for node != nil {
-		if scope := c.typeInfo.Scopes[node]; scope != nil {
-			return scope
-		}
-		switch node := node.(type) {
-		case *ast.FuncDecl:
-			if scope := c.typeInfo.Scopes[node.Type]; scope != nil {
-				return scope
-			}
-		case *ast.FuncLit:
-			if scope := c.typeInfo.Scopes[node.Type]; scope != nil {
-				return scope
-			}
-			if scope := c.typeInfo.Scopes[node.Body]; scope != nil {
-				return scope
-			}
-		}
-		node = c.parents[node]
-	}
-	return nil
-}
-
-// objectsInScope returns the objects in scope in scope-name order.
-func (c *inputSlotContext) objectsInScope(scope *gotypes.Scope) []gotypes.Object {
-	if objects, ok := c.scopeObjects[scope]; ok {
-		return objects
-	}
-	names := scope.Names()
-	objects := make([]gotypes.Object, 0, len(names))
-	for _, name := range names {
-		if obj := scope.Lookup(name); obj != nil {
-			objects = append(objects, obj)
-		}
-	}
-	c.scopeObjects[scope] = objects
-	return objects
-}
-
-// visibleObjectCount returns how many variables and constants in scope are
-// declared before pos.
-func (c *inputSlotContext) visibleObjectCount(scope *gotypes.Scope, pos token.Pos) int {
-	positions, ok := c.scopeObjectPositions[scope]
-	if !ok {
-		for _, obj := range c.objectsInScope(scope) {
-			switch obj.(type) {
-			case *gotypes.Var, *gotypes.Const:
-				positions = append(positions, obj.Pos())
-			}
-		}
-		slices.Sort(positions)
-		c.scopeObjectPositions[scope] = positions
-	}
-	count, _ := slices.BinarySearch(positions, pos)
-	return count
-}
-
-// compareInputSlotPriority compares input slots by conflict priority. Slots
-// with earlier starts win, and longer slots win when their starts are equal.
-func compareInputSlotPriority(a, b XGoInputSlot) int {
-	if start := comparePositions(a.Range.Start, b.Range.Start); start != 0 {
-		return start
-	}
-	return comparePositions(b.Range.End, a.Range.End)
-}
-
-// normalizeInputSlots orders input slots by source range, drops degenerate
-// ranges, and removes overlaps. Discovery order breaks ties between slots with
-// identical ranges. It takes ownership of inputSlots and may overwrite it.
-func normalizeInputSlots(inputSlots []XGoInputSlot) []XGoInputSlot {
-	// Reuse the input storage when candidates are already in priority order.
-	if slices.IsSortedFunc(inputSlots, compareInputSlotPriority) {
-		normalizedSlots := appendNonOverlappingInputSlots(inputSlots[:0], slices.Values(inputSlots))
-		clear(inputSlots[len(normalizedSlots):])
-		return normalizedSlots
-	}
-
-	// Sorting indices preserves discovery order for exact ties without moving
-	// the comparatively large slot values during every sort operation.
-	indices := make([]int, len(inputSlots))
-	for i := range indices {
-		indices[i] = i
-	}
-	slices.SortFunc(indices, func(i, j int) int {
-		if priority := compareInputSlotPriority(inputSlots[i], inputSlots[j]); priority != 0 {
-			return priority
-		}
-		return cmp.Compare(i, j)
-	})
-	orderedSlots := func(yield func(XGoInputSlot) bool) {
-		for _, index := range indices {
-			if !yield(inputSlots[index]) {
-				return
-			}
-		}
-	}
-	return appendNonOverlappingInputSlots(make([]XGoInputSlot, 0, len(inputSlots)), orderedSlots)
-}
-
-// appendNonOverlappingInputSlots appends valid slots from a sequence ordered by
-// [compareInputSlotPriority] unless they overlap the last accepted slot. The
-// ordering and disjointness of accepted slots mean only the last can overlap
-// the next slot. The destination must be empty and may alias storage read by
-// slots. Appending at most once per input keeps aliased writes behind unread
-// inputs.
-func appendNonOverlappingInputSlots(dst []XGoInputSlot, slots iter.Seq[XGoInputSlot]) []XGoInputSlot {
-	for slot := range slots {
-		if comparePositions(slot.Range.Start, slot.Range.End) >= 0 {
-			continue
-		}
-		if len(dst) > 0 && IsRangesOverlap(dst[len(dst)-1].Range, slot.Range) {
-			continue
-		}
-		dst = append(dst, slot)
-	}
-	return dst
-}
-
-// findInputSlots finds all input slots in the AST file in source order. The
-// returned slots have pairwise non-overlapping ranges.
-func findInputSlots(result *compileResult, astFile *ast.File) []XGoInputSlot {
-	typeInfo, _ := result.proj.TypeInfo()
-	if typeInfo == nil {
-		return nil
-	}
-	ctx := newInputSlotContext(result, astFile)
-
-	var inputSlots []XGoInputSlot
-	addInputSlot := func(slot *SpxInputSlot) {
-		if slot != nil {
-			inputSlots = append(inputSlots, *slot)
-		}
-	}
-
-	ast.Inspect(astFile, func(node ast.Node) bool {
-		if node == nil {
-			return true
-		}
-
-		switch node := node.(type) {
-		case *ast.BranchStmt:
-			if callExpr := xgoutil.CreateCallExprFromBranchStmt(typeInfo, node); callExpr != nil {
-				inputSlots = append(inputSlots, findInputSlotsFromCallExpr(ctx, callExpr)...)
-			}
-		case *ast.CallExpr, *ast.FuncDecorator:
-			inputSlots = append(inputSlots, findInputSlotsFromCallExpr(ctx, callExprFromNode(node))...)
-		case *ast.BinaryExpr:
-			addInputSlot(checkValueInputSlot(ctx, node.X, nil))
-			addInputSlot(checkValueInputSlot(ctx, node.Y, nil))
-		case *ast.UnaryExpr:
-			addInputSlot(checkValueInputSlot(ctx, node.X, nil))
-		case *ast.AssignStmt:
-			for _, lhs := range node.Lhs {
-				addInputSlot(checkAddressInputSlot(ctx, lhs))
-			}
-
-			for i, rhs := range node.Rhs {
-				var declaredType gotypes.Type
-				if len(node.Lhs) == len(node.Rhs) {
-					declaredType = typeInfo.TypeOf(node.Lhs[i])
-				}
-
-				addInputSlot(checkValueInputSlot(ctx, rhs, declaredType))
-			}
-		case *ast.ForStmt:
-			if expr, ok := node.Init.(*ast.ExprStmt); ok {
-				addInputSlot(checkValueInputSlot(ctx, expr.X, nil))
-			}
-
-			if node.Cond != nil {
-				addInputSlot(checkValueInputSlot(ctx, node.Cond, gotypes.Typ[gotypes.Bool]))
-			}
-
-			if expr, ok := node.Post.(*ast.ExprStmt); ok {
-				addInputSlot(checkValueInputSlot(ctx, expr.X, nil))
-			}
-		case *ast.ValueSpec:
-			for i, value := range node.Values {
-				var declaredType gotypes.Type
-				if len(node.Names) == len(node.Values) {
-					nameIdent := node.Names[i]
-					if nameIdent != nil && nameIdent.Name != "_" {
-						obj := typeInfo.ObjectOf(nameIdent)
-						if obj != nil {
-							declaredType = obj.Type()
-						}
-					}
-				}
-
-				addInputSlot(checkValueInputSlot(ctx, value, declaredType))
-			}
-		case *ast.ReturnStmt:
-			for _, res := range node.Results {
-				addInputSlot(checkValueInputSlot(ctx, res, nil))
-			}
-		case *ast.IfStmt:
-			addInputSlot(checkValueInputSlot(ctx, node.Cond, gotypes.Typ[gotypes.Bool]))
-		case *ast.SwitchStmt:
-			if node.Tag != nil {
-				addInputSlot(checkValueInputSlot(ctx, node.Tag, nil))
-			}
-		case *ast.CaseClause:
-			for _, expr := range node.List {
-				addInputSlot(checkValueInputSlot(ctx, expr, nil))
-			}
-		case *ast.RangeStmt:
-			if node.Key != nil && !isBlank(node.Key) {
-				addInputSlot(checkAddressInputSlot(ctx, node.Key))
-			}
-
-			if node.Value != nil && !isBlank(node.Value) {
-				addInputSlot(checkAddressInputSlot(ctx, node.Value))
-			}
-
-			addInputSlot(checkValueInputSlot(ctx, node.X, nil))
-		case *ast.IncDecStmt:
-			addInputSlot(checkAddressInputSlot(ctx, node.X))
-		}
-		return true
-	})
-	return normalizeInputSlots(inputSlots)
-}
-
-// findInputSlotsFromCallExpr finds input slots from a call expression.
-func findInputSlotsFromCallExpr(ctx *inputSlotContext, callExpr *ast.CallExpr) []SpxInputSlot {
-	if ctx.typeInfo == nil {
-		return nil
-	}
-
-	var inputSlots []SpxInputSlot
-	for resolvedArg := range resolvedCallExprArgs(ctx.result.proj, ctx.typeInfo, callExpr) {
-		if resolvedArg.ExpectedType == nil || resolvedArg.IsTypeArg() {
-			continue
-		}
-
-		expectedType := resolvedArg.ExpectedType
-		if _, ok := expectedType.(*gotypes.TypeParam); ok {
-			if actualType := ctx.typeInfo.TypeOf(resolvedArg.Arg); xgoutil.IsValidType(actualType) {
-				expectedType = actualType
-			}
-		}
-		declaredType := xgoutil.DerefType(expectedType)
-		if sliceType, ok := declaredType.(*gotypes.Slice); ok {
-			declaredType = xgoutil.DerefType(sliceType.Elem())
-		}
-
-		var slot *SpxInputSlot
-		if lit, ok := resolvedArg.Arg.(*ast.NumberUnitLit); ok {
-			unitExpectedType := xgoUnitExpectedTypeForResolvedArg(resolvedArg)
-			if len(xgoUnitSpecsForType(unitExpectedType)) == 0 {
-				continue
-			}
-			declaredType = xgoutil.DerefType(unitExpectedType)
-			slot = createValueInputSlotFromNumberUnitLit(ctx, lit, declaredType)
-		} else {
-			slot = checkValueInputSlot(ctx, resolvedArg.Arg, declaredType)
-		}
-		if slot != nil {
-			inputSlots = append(inputSlots, *slot)
-		}
-	}
-	return inputSlots
-}
-
-// collectPredefinedNames collects all predefined names for the given expression.
-func collectPredefinedNames(ctx *inputSlotContext, expr ast.Expr, declaredType gotypes.Type) []string {
-	innermostScope := ctx.innermostScope(expr)
-	key := predefinedNamesCacheKey{scope: innermostScope, declaredType: declaredType}
-	if innermostScope != nil {
-		key.visibleObjectCount = ctx.visibleObjectCount(innermostScope, expr.Pos())
-	}
-	if names, ok := ctx.predefinedNames[key]; ok {
-		return names
-	}
-
-	var names []string
-	seenNames := make(map[string]struct{})
-	addObjectName := func(obj gotypes.Object) {
-		name := obj.Name()
-		switch obj := obj.(type) {
-		case *gotypes.Var, *gotypes.Const:
-			if typ := obj.Type(); typ != nil && declaredType != nil && !gotypes.AssignableTo(typ, declaredType) {
-				return
-			}
-
-			if name == "this" || xgoutil.IsXGoInternalName(name) {
-				return
-			}
-		case *gotypes.Func:
-			if declaredType != nil {
-				// For functions with no parameters and exactly one return value,
-				// check if the return type is assignable to the declared type.
-				funcSig := obj.Signature()
-				if funcSig.Params().Len() != 0 || funcSig.Results().Len() != 1 {
-					return
-				}
-				funcReturnType := funcSig.Results().At(0).Type()
-				if !gotypes.AssignableTo(funcReturnType, declaredType) {
-					return
-				}
-			}
-
-			name = xgoutil.ToLowerCamelCase(name)
-		default:
-			return
-		}
-		if _, ok := seenNames[name]; ok {
-			return
-		}
-		seenNames[name] = struct{}{}
-		names = append(names, name)
-	}
-
-	for scope := innermostScope; scope != nil && scope != gotypes.Universe; scope = scope.Parent() {
-		objects := ctx.objectsInScope(scope)
-		names = slices.Grow(names, len(objects))
-		for _, obj := range objects {
-			if scope != innermostScope || obj.Pos() < expr.Pos() {
-				switch obj.(type) {
-				case *gotypes.Var, *gotypes.Const:
-					addObjectName(obj)
-				}
-			}
-
-			if !ctx.astFile.IsClass || !xgoutil.IsSyntheticThisIdent(
-				ctx.result.proj.Fset,
-				ctx.typeInfo,
-				ctx.astPkg,
-				ctx.typeInfo.ObjToDef[obj],
-			) {
-				continue
-			}
-			objType := xgoutil.DerefType(obj.Type())
-			named, ok := objType.(*gotypes.Named)
-			if !ok || !xgoutil.IsNamedStructType(named) {
-				continue
-			}
-
-			for structMember := range xgoutil.StructMembers(named) {
-				switch member := structMember.Member.(type) {
-				case *gotypes.Var:
-					if !member.Origin().Embedded() {
-						addObjectName(member)
-					}
-				case *gotypes.Func:
-					// Add methods with no parameters and exactly one return value.
-					// For example, the method `Game.BackdropName` can be used in `echo backdropname`.
-					funcSig := member.Signature()
-					if funcSig.Params().Len() == 0 && funcSig.Results().Len() == 1 {
-						addObjectName(member)
-					}
-				}
-			}
-		}
-	}
-
-	for _, scope := range []*gotypes.Scope{
-		GetSpxPkg().Scope(),
-		GetMathPkg().Scope(),
-		gotypes.Universe,
-	} {
-		objects := ctx.objectsInScope(scope)
-		names = slices.Grow(names, len(objects))
-		for _, obj := range objects {
-			if _, ok := obj.(*gotypes.Var); ok {
-				addObjectName(obj)
-			}
-		}
-	}
-
-	ctx.predefinedNames[key] = names
-	return names
-}
-
-// checkValueInputSlot checks if the expression is a value input slot.
-func checkValueInputSlot(ctx *inputSlotContext, expr ast.Expr, declaredType gotypes.Type) *SpxInputSlot {
-	switch expr := expr.(type) {
-	case *ast.BasicLit:
-		return createValueInputSlotFromBasicLit(ctx, expr, declaredType)
-	case *ast.Ident:
-		return createValueInputSlotFromIdent(ctx, expr, declaredType)
-	case *ast.UnaryExpr:
-		return createValueInputSlotFromUnaryExpr(ctx, expr, declaredType)
-	case *ast.CallExpr:
-		return createValueInputSlotFromColorFuncCall(ctx, expr, declaredType)
-	}
-	return nil
-}
-
-// checkAddressInputSlot checks if the expression is an address input slot.
-func checkAddressInputSlot(ctx *inputSlotContext, expr ast.Expr) *SpxInputSlot {
-	ident, ok := expr.(*ast.Ident)
-	if !ok {
-		return nil
-	}
-	return &SpxInputSlot{
-		Kind:   SpxInputSlotKindAddress,
-		Accept: SpxInputSlotAccept{Type: SpxInputTypeUnknown},
-		Input: SpxInput{
-			Kind: SpxInputKindPredefined,
-			Type: SpxInputTypeUnknown,
-			Name: ident.Name,
-		},
-		PredefinedNames: collectPredefinedNames(ctx, expr, nil),
-		Range:           ctx.rangeForNode(ident),
-	}
-}
-
-// createValueInputSlotFromBasicLit creates a value input slot from a basic literal.
-func createValueInputSlotFromBasicLit(ctx *inputSlotContext, lit *ast.BasicLit, declaredType gotypes.Type) *SpxInputSlot {
-	input := SpxInput{Kind: SpxInputKindInPlace}
-	switch lit.Kind {
-	case token.STRING:
-		input.Type = SpxInputTypeString
-		v, err := strconv.Unquote(lit.Value)
-		if err != nil {
-			return nil
-		}
-		input.Value = v
-	case token.INT:
-		input.Type = SpxInputTypeInteger
-		v, err := strconv.ParseInt(lit.Value, 0, 64)
-		if err != nil {
-			return nil
-		}
-		input.Value = v
-	case token.FLOAT:
-		input.Type = SpxInputTypeDecimal
-		v, err := strconv.ParseFloat(lit.Value, 64)
-		if err != nil {
-			return nil
-		}
-		input.Value = v
-	default:
-		return nil
-	}
-
-	accept := SpxInputSlotAccept{Type: input.Type}
-	if declaredType != nil {
-		accept.Type = inferSpxInputTypeFromType(declaredType)
-	}
-	if accept.Type == SpxInputTypeResourceName {
-		for _, spxResourceRef := range ctx.result.spxResourceRefs {
-			if spxResourceRef.Node == lit {
-				input.Type = SpxInputTypeResourceName
-				input.Value = spxResourceRef.ID.URI()
-				accept.ResourceContext = ToPtr(spxResourceRef.ID.ContextURI())
-				break
-			}
-		}
-		if accept.ResourceContext == nil {
-			return nil
-		}
-	}
-
-	return &SpxInputSlot{
-		Kind:            SpxInputSlotKindValue,
-		Accept:          accept,
-		Input:           input,
-		PredefinedNames: collectPredefinedNames(ctx, lit, declaredType),
-		Range:           ctx.rangeForNode(lit),
-	}
-}
-
-// createValueInputSlotFromNumberUnitLit creates a value input slot from a
-// number-with-unit literal.
-func createValueInputSlotFromNumberUnitLit(ctx *inputSlotContext, lit *ast.NumberUnitLit, declaredType gotypes.Type) *SpxInputSlot {
-	input := SpxInput{Kind: SpxInputKindInPlace}
-	switch lit.Kind {
-	case token.INT:
-		input.Type = SpxInputTypeInteger
-		v, err := strconv.ParseInt(lit.Value, 0, 64)
-		if err != nil {
-			return nil
-		}
-		input.Value = v
-	case token.FLOAT:
-		input.Type = SpxInputTypeDecimal
-		v, err := strconv.ParseFloat(lit.Value, 64)
-		if err != nil {
-			return nil
-		}
-		input.Value = v
-	default:
-		return nil
-	}
-
-	accept := SpxInputSlotAccept{Type: input.Type}
-	if declaredType != nil {
-		if acceptType := inferSpxInputTypeFromType(declaredType); acceptType != SpxInputTypeUnknown {
-			accept.Type = acceptType
-		}
-	}
-
-	return &SpxInputSlot{
-		Kind:            SpxInputSlotKindValue,
-		Accept:          accept,
-		Input:           input,
-		PredefinedNames: collectPredefinedNames(ctx, lit, declaredType),
-		Range:           ctx.rangeForPosEnd(lit.ValuePos, xgoUnitStart(lit)),
-	}
-}
-
-// createValueInputSlotFromIdent creates a value input slot from an identifier.
-func createValueInputSlotFromIdent(ctx *inputSlotContext, ident *ast.Ident, declaredType gotypes.Type) *SpxInputSlot {
-	if ctx.typeInfo == nil {
-		return nil
-	}
-	typ := ctx.typeInfo.TypeOf(ident)
-	if typ == nil {
-		return nil
-	}
-	typ = xgoutil.DerefType(typ)
-
-	input := SpxInput{
-		Kind: SpxInputKindPredefined,
-		Type: inferSpxInputTypeFromTypeInProject(ctx.result, typ),
-		Name: ident.Name,
-	}
-	switch input.Type {
-	case SpxInputTypeBoolean:
-		if basicType, ok := typ.(*gotypes.Basic); ok && basicType.Kind() == gotypes.UntypedBool {
-			input.Kind = SpxInputKindInPlace
-			input.Value = ident.Name == "true"
-			input.Name = ""
-		}
-	case SpxInputTypeDirection,
-		SpxInputTypeEffectKind,
-		SpxInputTypeLayerAction,
-		SpxInputTypeDirAction,
-		SpxInputTypeKey,
-		SpxInputTypeSpecialObj,
-		SpxInputTypeRotationStyle:
-		obj := ctx.typeInfo.ObjectOf(ident)
-		if obj != nil && !IsInSpxPkg(obj) {
-			break
-		}
-		cnst, ok := obj.(*gotypes.Const)
-		if !ok {
-			break
-		}
-		input.Kind = SpxInputKindInPlace
-		switch input.Type {
-		case SpxInputTypeDirection:
-			input.Value, _ = strconv.ParseFloat(cnst.Val().ExactString(), 64)
-		default:
-			input.Value = cnst.Name()
-		}
-		input.Name = ""
-	}
-
-	accept := SpxInputSlotAccept{Type: input.Type}
-	if declaredType != nil {
-		accept.Type = inferSpxInputTypeFromTypeInProject(ctx.result, declaredType)
-	}
-	switch accept.Type {
-	case SpxInputTypeResourceName:
-		switch canonicalSpxResourceNameType(declaredType) {
-		case GetSpxBackdropNameType():
-			accept.ResourceContext = ToPtr(SpxBackdropResourceContextURI)
-		case GetSpxSoundNameType():
-			accept.ResourceContext = ToPtr(SpxSoundResourceContextURI)
-		case GetSpxSpriteNameType():
-			accept.ResourceContext = ToPtr(SpxSpriteResourceContextURI)
-		case GetSpxSpriteCostumeNameType():
-			spxSpriteResource := inferSpxSpriteResourceEnclosingNode(ctx.result, ident)
-			if spxSpriteResource == nil {
-				return nil
-			}
-			accept.ResourceContext = ToPtr(FormatSpxSpriteCostumeResourceContextURI(spxSpriteResource.Name))
-		case GetSpxSpriteAnimationNameType():
-			spxSpriteResource := inferSpxSpriteResourceEnclosingNode(ctx.result, ident)
-			if spxSpriteResource == nil {
-				return nil
-			}
-			accept.ResourceContext = ToPtr(FormatSpxSpriteAnimationResourceContextURI(spxSpriteResource.Name))
-		case GetSpxWidgetNameType():
-			accept.ResourceContext = ToPtr(SpxWidgetResourceContextURI)
-		default:
-			return nil
-		}
-	case SpxInputTypeSpriteInstance:
-		accept.ResourceContext = ToPtr(SpxSpriteResourceContextURI)
-		if spxSpriteResource := spxSpriteResourceForObject(ctx.result, ctx.typeInfo.ObjectOf(ident)); spxSpriteResource != nil {
-			input.Kind = SpxInputKindInPlace
-			input.Value = spxSpriteResource.ID.URI()
-			input.Name = ""
-		}
-	}
-
-	return &SpxInputSlot{
-		Kind:            SpxInputSlotKindValue,
-		Accept:          accept,
-		Input:           input,
-		PredefinedNames: collectPredefinedNames(ctx, ident, declaredType),
-		Range:           ctx.rangeForNode(ident),
-	}
-}
-
-// inferSpxInputTypeFromTypeInProject attempts to infer the input type from typ
-// using project sprite type metadata.
-func inferSpxInputTypeFromTypeInProject(result *compileResult, typ gotypes.Type) SpxInputType {
-	if isSpxSpriteInstanceType(result, typ) {
-		return SpxInputTypeSpriteInstance
-	}
-	return inferSpxInputTypeFromType(typ)
-}
-
-// isSpxSpriteInstanceType reports whether the given type represents an spx
-// sprite instance.
-func isSpxSpriteInstanceType(result *compileResult, typ gotypes.Type) bool {
-	if typ == nil {
-		return false
-	}
-	typ = xgoutil.DerefType(typ)
-	if typ == GetSpxSpriteType() {
-		return true
-	}
-	if result != nil && result.hasSpxSpriteType(typ) {
-		return true
-	}
-	return gotypes.AssignableTo(typ, GetSpxSpriteType())
-}
-
-// spxSpriteResourceForObject returns the spx sprite resource for obj if it is an
-// auto-bound sprite. It returns nil if obj is nil, obj has no auto-binding, or
-// the corresponding sprite resource is not found in the resource set.
-func spxSpriteResourceForObject(result *compileResult, obj gotypes.Object) *SpxSpriteResource {
-	if obj == nil || !result.hasSpxSpriteResourceAutoBinding(obj) {
-		return nil
-	}
-	return result.spxResourceSet.Sprite(obj.Name())
-}
-
-// createValueInputSlotFromUnaryExpr creates a value input slot from a unary expression.
-func createValueInputSlotFromUnaryExpr(ctx *inputSlotContext, expr *ast.UnaryExpr, declaredType gotypes.Type) *SpxInputSlot {
-	var inputSlot *SpxInputSlot
-	switch x := expr.X.(type) {
-	case *ast.BasicLit:
-		inputSlot = createValueInputSlotFromBasicLit(ctx, x, declaredType)
-		if inputSlot == nil {
-			return nil
-		}
-
-		switch expr.Op {
-		case token.ADD:
-			// Nothing to do for unary plus.
-		case token.SUB:
-			switch v := inputSlot.Input.Value.(type) {
-			case int64:
-				inputSlot.Input.Value = -v
-			case float64:
-				inputSlot.Input.Value = -v
-			default:
-				return nil
-			}
-		case token.XOR:
-			switch x.Kind {
-			case token.INT:
-				switch v := inputSlot.Input.Value.(type) {
-				case int64:
-					inputSlot.Input.Value = ^v
-				default:
-					return nil
-				}
-			default:
-				return nil
-			}
-		}
-	case *ast.Ident:
-		inputSlot = createValueInputSlotFromIdent(ctx, x, declaredType)
-		if inputSlot == nil {
-			return nil
-		}
-
-		switch expr.Op {
-		case token.NOT:
-			switch v := inputSlot.Input.Value.(type) {
-			case bool:
-				inputSlot.Input.Value = !v
-			default:
-				return nil
-			}
-		}
-	default:
-		return nil
-	}
-	inputSlot.Range = ctx.rangeForNode(expr) // Update the range to include the entire unary expression.
-	return inputSlot
-}
-
-// createValueInputSlotFromColorFuncCall creates a value input slot from an spx
-// color function call.
-func createValueInputSlotFromColorFuncCall(ctx *inputSlotContext, callExpr *ast.CallExpr, declaredType gotypes.Type) *SpxInputSlot {
-	if ctx.typeInfo == nil {
-		return nil
-	}
-
-	fun := xgoutil.FuncFromCallExpr(ctx.typeInfo, callExpr)
-	if fun == nil || !IsInSpxPkg(fun) || !isSpxColorFunc(fun) {
-		return nil
-	}
-
-	constructor := SpxInputTypeSpxColorConstructor(fun.Name())
-	maxArgs := 3
-	switch constructor {
-	case SpxInputTypeSpxColorConstructorHSB:
-	case SpxInputTypeSpxColorConstructorHSBA:
-		maxArgs = 4
-	default:
-		return nil // This should never happen, but just in case.
-	}
-
-	var args []float64
-	for i, argExpr := range callExpr.Args {
-		if i >= maxArgs {
-			break
-		}
-		lit, ok := argExpr.(*ast.BasicLit)
-		if !ok {
-			return nil
-		}
-
-		var val float64
-		switch lit.Kind {
-		case token.FLOAT:
-			floatVal, err := strconv.ParseFloat(lit.Value, 64)
-			if err != nil {
-				return nil
-			}
-			val = floatVal
-		case token.INT:
-			intVal, err := strconv.ParseInt(lit.Value, 0, 64)
-			if err != nil {
-				return nil
-			}
-			val = float64(intVal)
-		default:
-			return nil
-		}
-		args = append(args, val)
-	}
-	if len(args) < maxArgs {
-		return nil
-	}
-
-	return &SpxInputSlot{
-		Kind:   SpxInputSlotKindValue,
-		Accept: SpxInputSlotAccept{Type: SpxInputTypeColor},
-		Input: SpxInput{
-			Kind: SpxInputKindInPlace,
-			Type: SpxInputTypeColor,
-			Value: SpxColorInputValue{
-				Constructor: constructor,
-				Args:        args,
-			},
-		},
-		PredefinedNames: collectPredefinedNames(ctx, callExpr, declaredType),
-		Range:           ctx.rangeForNode(callExpr),
-	}
-}
-
-// isSpxColorFunc checks if the fun is an spx color function.
-func isSpxColorFunc(fun *gotypes.Func) bool {
-	switch fun {
-	case GetSpxHSBFunc(), GetSpxHSBAFunc():
-		return true
-	}
-	return false
-}
-
-// inferSpxInputTypeFromType attempts to infer the input type from the given type.
-func inferSpxInputTypeFromType(typ gotypes.Type) SpxInputType {
-	if basicType, ok := typ.(*gotypes.Basic); ok {
-		switch basicType.Kind() {
-		case gotypes.String, gotypes.UntypedString:
-			return SpxInputTypeString
-		case gotypes.Int, gotypes.Int8, gotypes.Int16, gotypes.Int32, gotypes.Int64,
-			gotypes.Uint, gotypes.Uint8, gotypes.Uint16, gotypes.Uint32, gotypes.Uint64,
-			gotypes.UntypedInt:
-			return SpxInputTypeInteger
-		case gotypes.Float32, gotypes.Float64, gotypes.UntypedFloat:
-			return SpxInputTypeDecimal
-		case gotypes.Bool, gotypes.UntypedBool:
-			return SpxInputTypeBoolean
-		}
-		return SpxInputTypeUnknown
-	}
-
-	if IsSpxResourceNameType(typ) {
-		return SpxInputTypeResourceName
-	}
-
-	switch typ {
-	case GetSpxDirectionType():
-		return SpxInputTypeDirection
-	case GetSpxLayerActionType():
-		return SpxInputTypeLayerAction
-	case GetSpxDirActionType():
-		return SpxInputTypeDirAction
-	case GetSpxEffectKindType():
-		return SpxInputTypeEffectKind
-	case GetSpxKeyType():
-		return SpxInputTypeKey
-	case GetSpxSpecialObjType():
-		return SpxInputTypeSpecialObj
-	case GetSpxRotationStyleType():
-		return SpxInputTypeRotationStyle
-	case GetSpxPropertyNameType():
-		return SpxInputTypePropertyName
-	}
-
-	// Fall back to the alias RHS when no direct basic or spx type match is found.
-	if alias, ok := typ.(*gotypes.Alias); ok {
-		rhs := alias.Rhs()
-		if rhs != nil && rhs != typ {
-			return inferSpxInputTypeFromType(rhs)
-		}
-	}
-	return SpxInputTypeUnknown
-}
-
-// inferSpxSpriteResourceEnclosingNode infers the enclosing [SpxSpriteResource]
-// for the given node. It returns nil if no [SpxSpriteResource] can be inferred.
-func inferSpxSpriteResourceEnclosingNode(result *compileResult, node ast.Node) *SpxSpriteResource {
-	typeInfo, _ := result.proj.TypeInfo()
-	if typeInfo == nil {
-		return nil
-	}
-	spxFile := xgoutil.NodeFilename(result.proj.Fset, node)
-	astPkg, _ := result.proj.ASTPackage()
-	astFile := xgoutil.NodeASTFile(result.proj.Fset, astPkg, node)
-
-	var spxSpriteResource *SpxSpriteResource
-	for pathNode := range xgoutil.PathEnclosingIntervalNodes(astFile, node.Pos(), node.End(), false) {
-		if pathNode == nil {
-			continue
-		}
-
-		callExpr := callExprFromNode(pathNode)
-		if callExpr == nil {
-			continue
-		}
-
-		var spxSpriteName string
-		if sel, ok := callExpr.Fun.(*ast.SelectorExpr); ok {
-			ident, ok := sel.X.(*ast.Ident)
-			if !ok {
-				break
-			}
-			obj := typeInfo.ObjectOf(ident)
-			if obj == nil {
-				break
-			}
-			named, ok := xgoutil.DerefType(obj.Type()).(*gotypes.Named)
-			if !ok {
-				break
-			}
-
-			if named == GetSpxSpriteType() {
-				spxSpriteName = ident.Name
-			} else if result.hasSpxSpriteType(named) {
-				spxSpriteName = obj.Name()
-			}
-		} else if spxFile != "main.spx" {
-			spxSpriteName = strings.TrimSuffix(spxFile, ".spx")
-		}
-		spxSpriteResource = result.spxResourceSet.sprites[spxSpriteName]
-		break
-	}
-	return spxSpriteResource
-}
-
-// isBlank checks if an expression is a blank identifier (_).
-func isBlank(expr ast.Expr) bool {
-	ident, ok := expr.(*ast.Ident)
-	return ok && ident.Name == "_"
 }
